requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
chartjs==1.2
crispy-tailwind==0.5.0
Django==4.1.2
django-browser-reload==1.6.0
django-crispy-forms==1.14.0
django-storages==1.13.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
#mysqlclient==2.1.1
Pillow==9.2.0
psycopg2-binary==2.9.5
pytz==2022.6
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.12
whitenoise==6.2.0
